requirements.txt
pandas==2.2.3
matplotlib==3.9.4
openpyxl==3.1.5
numpy==2.2.3
python-telegram-bot==21.11